warehouse: [refactor] Create warehouse directory
